{{cookiecutter.bot_name}}/requirements.txt
botbuilder-core>=4.7.0
aiohttp

coco-microsoft-bot-framework